Add tests for guest redirects
This moves the setup code into the ConnCase as a new authenticate/1
function which we instead call via `set [:authenticate]` when needed.

With the auth setup now optional, we're able to add tests for the routes
that make sure guests are redirected to the login page. This couldn't be
done before because the global setup meant *all* conns were authed.