Improve test authentication
Rather than have a separate authenticate/1 function which needs to be
explicitly included in the setup process, this switches to a "parent"
setup/1 function to run it automatically. This is possible now because
rather than mutating the primary `conn`, we're returning an extra authed
connection in addition to the default conn that's normally returned.

Another benefit of this way is not only can we merge our guest tests
into the main describe blocks, but we also no longer need to worry about
overriding the connection and saving it as `resp` or similar instead.

tl;dr: We can now use authed+guest connections simultaneously!