Add support for roles
This adds three roles: admin, moderator and unverified for newly
registered users that need to be manually "promoted" to moderator.

Also adds the necessary pipelines and updates our router to enforce the
respective roles for mod/admin routes. Both mods and admins will see the
same dashboard route, so that one's a bit more lax than the rest.