Add backend UI for managing users
Adds the context, schema, controller, and frontend components needed in
order to manage the internal users from the backend. To achieve this
without affecting Pow's default routes such as editing your own account,
we've had to add an extra `admin_changeset` to use from the main schema.

Also renames the "Admin" section of the sidebar to allow for an extra.