Set page titles based on controller assigns
Adds a new `page_title/1` helper function to the layout view which takes
the main title passed in from the controller and spits out the full page
title for use in the `<title>` tag in the main app layout template.

Since we're setting the title in our templates anyway so that it can be
passed to the layout, we also use it in the visible page titles.