Add basic email domain listing
Adds a new controller/view that spits out a list of all domain names
used in the email address of any forum members, including a count of how
many members are using that domain.

Could benefit from a whitelist (eg gmail.com) and blacklist (eg *.xyz),
but this is good enough to get an overview of the most popular ones.