Add bans associated with an email domain
Adds a new Ecto schema for the banlist table so that we can use it to
join into the email domain query, which is in turn tweaked slightly to
use proper table names from our schema in the select rather than
hard-coding the field name directly.

For now we only show the value of the `ban_email` field, which will be
something along the lines of `*@example.com` if it's set. This'll need
to be replaced with a sort of "ban status" that eventually also checks
whether or not any ban is still active or has already expired.