Add basic whitelist indications
Purely visual for the moment, but this adds a short list of somewhat
trusted email providers which can later be used to prevent new global
bans being applied to them. Unless of course a matching ban already
exists for that particular email domain, in which case the blacklist
will take priority.

Just in case, we also add `true` as a fallback if somehow none of the
other options in our condition end up matching to any of the handlers.