Add ban status to email domain details page
All we previously had available on the actual detail page was the domain
name itself, as a string. That's not gonna work for our ban test because
that simply casts to bool assuming it's given a struct conforming to
`%Ban{}`, ie one providing the `email_domain` field.

To get around that, we add a new query to the forum context which will
grab the first matching entry from the banlist table, based on the same
filter that would be used when selecting members by email domain.