Fix ban status and add more detail
This adds handling for `nil` ban results on the details page such that
the whitelist is still checked and errors are prevented when attempting
to display fields from the struct that doesn't exist in the first place.

We also add some extra context to the status output so it's clear how a
particular domain is blacklisted, if at all.

Note this uses what is probably a bit of a hack to inject the whitelist
into the template so we can check it there as well as within the view.
There's likely a proper way to do that, but I don't know it yet.