Add flash data when blacklisting an email domain
This will not work without piping unless we assign the result of
`put_flash/2` to a new `conn` variable. In elixir, it's all immutable
and thus the changes to `conn` are lost if we simply pass the same
`conn` into redirect without saving the new one from `put_flash`.

It works with piping because the result of one function is automatically
passed on to the next. Thanks to Nicd- in #elixir-lang for the explain!