Get (almost) everything converted to new layouts
Three days in the making thanks to a bunch of irl distractions, but this
finally converts the user, member and email domain views/templates to
the new-style component-based layouts and controllers.

Live dashboard link in the sidebar still isn't working, and nor is
pagination with the scrivener html/phoenix lib, so there's more playing
with forks to figure that out, but everything else seems to be working.